Move static methods to TaskManagerInstances class
Finally understood that we are tracking the in-flight resource
consumption of instances via SimpleNamespace objects we are carrying
around in ig_capacity_graph[ig_name][instances]. Moved the static
methods that choose what the best one of these is into the
TaskManagerInstances class so all the logic that determines what
instance we choose and what the in-flight capacity consumption is all
under one roof.
